+++ src/main/kotlin/com/github/denofevil/aurelia/completion/AureliaAttributeCompletionProvider.kt
@@ -1,6 +1,7 @@
 package com.github.denofevil.aurelia.completion
 
 import com.github.denofevil.aurelia.Aurelia
+import com.github.denofevil.aurelia.attribute.AttributeUtil
 import com.github.denofevil.aurelia.index.AureliaIndexUtil
 import com.github.denofevil.aurelia.require.DeclarationResolverUtil
 import com.intellij.codeInsight.completion.CompletionParameters
@@ -20,11 +21,13 @@ class AureliaAttributeCompletionProvider : CompletionProvider<CompletionParamete
         val tag = element.parent ?: return
 
         AureliaIndexUtil.getAllCustomAttributeNames(tag.project).forEach {
+            if (tag.attributes.any { atr -> AttributeUtil.withoutInjectable(atr.name) == it }) return@forEach
             result.addElement(LookupElementBuilder.create(it).withIcon(Aurelia.ATTRIBUTE_ICON))
         }
         val jsClass = DeclarationResolverUtil.resolveCustomElementDeclaration(tag)
         DeclarationResolverUtil.resolveBindableAttributesOnlyWithAnnotation(jsClass).forEach {
             val text = Aurelia.camelToKebabCase(it.memberName).lowercase()
+            if (tag.attributes.any { atr -> AttributeUtil.withoutInjectable(atr.name) == text }) return@forEach
             result.addElement(LookupElementBuilder.create(text).withIcon(Aurelia.PROPERTY_ICON))
         }
     }