+++ CHANGELOG.md
@@ -9,8 +9,10 @@
 - Proper `repeat.for` detection and reference resolving
 - Ignoring binding behaviors and value converts for code injection
 - Suppressing of missing promise for aurelia js injected code
-- $this, $parent, $index, $event support for code injections
+- `$this`, `$parent`, `$index`, `$event` support for code injections
 - GoTo Declaration/definition (Default Alt+Ctrl+Shift+O)
+- Custom element completion (ctrl+space)
+- Property and custom attribute completion (ctrl+space)
 
 ## [2.1.0]